requirements.txt
beautifulsoup4==4.11.2
Flask==2.2.2
gensim==4.3.0
python_Levenshtein==0.20.9
requests==2.25.1
selenium==4.8.0
sentence_transformers==2.2.2
